refactor(api): move sendInvitations method from organization-controller
to src/team/organization-invitations controller
